Use the CLEARDB_DATABASE_URL for Heroku
This should make launching on Heroku easier.

We're telling people to install ClearDB, let's just use the URL that works with that. I spent about 30 mins trying to figure out why MySQL wasn't connecting and then realised it's because the CLEARDB_DATABASE_URL doesn't have a port, they just use the default port. So now if there is no port specified it doesn't put it on the end of the URL.